[BUGFIX] Make sure TS generation uses current pid in BE context
This might fix issues with wrong TS in BE context (e.g. when comments are approved in BE).
The fix helps to make sure pid passed to FE link action VH is correct (fixes wrong links
in multi domain and sub folder setups) and ensures proper TS in email in (mutli domain
setups)

Related to #166 and #164