[!!!][TASK] Use new dependency injection configuration (incl. registering interfaces)
Make use of the new Symfony DI component for better performance and in order to make use of best practice.

Add constructor argument injection as inject annotations (but keep them in place for TYPO3 9.x for now)
won't work with new DI.

Interfaces should be configured for TYPO3 10 as TS configuration has been removed.
Not really needed here as we only have a single implementation for each class but configured
interfaces anyway.

https://docs.typo3.org/c/typo3/cms-core/master/en-us/Changelog/9.5/Deprecation-86270-ExtbaseXclassViaTypoScriptSettings.html
https://docs.typo3.org/c/typo3/cms-core/master/en-us/Changelog/10.0/Feature-84112-SymfonyDependencyInjectionForCoreAndExtbase.html
https://docs.typo3.org/c/typo3/cms-core/master/en-us/Changelog/9.0/Deprecation-82975-DeprecateUsageOfInjectWithNonPublicProperties.html
https://usetypo3.com/dependency-injection.html